% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotoneFIT.R
\name{monotoneFIT}
\alias{monotoneFIT}
\title{Extract Fitted values after Bayesian estimation of a Monotone Single Index Model.}
\usage{
monotoneFIT(mono.sim, size.grid.x = 100, grid.x = NULL)
}
\arguments{
\item{mono.sim}{A returned object of \code{monotoneSIM} function.}

\item{size.grid.x}{length of vector of values for which the monotone function is estimated. Takes the value \eqn{100} by default.}

\item{grid.x}{(Optional) Vector of user supplied values for which the monotone function is estimated. Takes \code{NULL} by default. If \code{NULL}, vector (of length size.grid.x) of equispaced values between -1 and 1 are chosen.}
}
\value{
A list with the following elements.
\item{Y.fitted}{ \code{M} \eqn{x n} Matrix of Fitted values of the response. \code{M} is the size of the MCMC sample obtained in \code{monotoneSIM}. }
\item{g.estimated}{ \code{M} \eqn{x} \code{size.grid.x} Matrix OR (if \code{grid.x} is supplied) \code{M} \eqn{x} \code{length(grid.x)} Matrix of estimated values of the link function. Each row represents estimated value of the link function evaluated on grid.x, corresponding to each posterior sample of basis coefficients.}
\item{grid.x}{ Grid of x values at which the link function is estimated. }
}
\description{
Extract Fitted values after Bayesian estimation of a Monotone Single Index Model.
}
