% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotoneSIM_wrapper.R
\name{monotoneSIM}
\alias{monotoneSIM}
\title{Bayesian Estimation of Monotone Single Index Models}
\usage{
monotoneSIM(
  y,
  X,
  beta.init,
  xi.init,
  Sigma.xi,
  knots,
  monotone = TRUE,
  iter.HMC = 10,
  sigma.sq.beta = 10,
  sigma.sq.eps = 1,
  a.eps = 1,
  b.eps = 1,
  Burn.in = 100,
  M = 1000,
  grid.x = NULL,
  size.grid.x = 100
)
}
\arguments{
\item{size.grid.x}{}
}
\value{

}
\description{
Bayesian Estimation of Monotone Single Index Models
}
